+++ controller/usercontroller.js
@@ -1,22 +1,27 @@
 var NEWUSER = require("../Model/user");
-var ADMIN = require("../Model/admin");
 const bycrypt = require('bcrypt');
 const jwt = require('jsonwebtoken');
+const fs = require('fs');
+const path = require('path');
+const multer = require('multer');
+const upload = multer({ dest: 'uploads/' }).single('profile'); // Temp storage
 
 
 exports.SECURE = async function (req, res, next) {
     try {
 
-    const token = req.headers.authorization;
-    if(!token) throw new Error("PLease Attche Token");
+        const token = req.headers.authorization;
+        if (!token) throw new Error("PLease Attche Token");
 
-    const isvalidtoken = jwt.verify(token,"TEST");
-    const  tokenndata = await ADMIN.findById(isvalidtoken.id);
-    
-    if(!tokenndata) throw new Error("USER is not valid");
-      
-    next()
-    
+        const isvalidtoken = jwt.verify(token, "USER");
+        const tokenndata = await NEWUSER.findById(isvalidtoken.id);
+
+        if (!tokenndata) throw new Error("USER is not valid");
+
+        req.user = tokenndata._id
+
+        next()
+
     } catch (error) {
         res.status(404).json({
             status: "Fail",
@@ -25,21 +30,37 @@ exports.SECURE = async function (req, res, next) {
     }
 }
 
+exports.MULTER = async function (req, res, next) {
+    try {
+
+
+
+        next()
+
+    } catch (error) {
+        res.status(404).json({
+            status: "Fail",
+            message: error.message,
+        })
+    }
+}
+
+
+
+
 exports.SIGNUP = async function (req, res, next) {
     try {
         // console.log(req.file)
         // return res.send("test")
-        
-        let { firstname, lastname, email, password } = req.body;
+        let { email, password, role } = req.body;
         password = await bycrypt.hash(password, 8);
 
         let usercreate = await NEWUSER.create({
-            firstname,
-            lastname,
-            email,
-            password,
-            profile: req.file.filename
-        })
+                email,
+                password,
+                profile:  req.file?.filename,
+                role
+            })
 
         res.status(201).json({
             status: "Success",
@@ -59,12 +80,11 @@ exports.SIGNUP = async function (req, res, next) {
 exports.LOGIN = async function (req, res, next) {
     try {
         let { email, password } = req.body;
-        const userfind = await NEWUSER.findOne({ email });
 
+        const userfind = await NEWUSER.findOne({ email });
         if (!userfind) throw new Error('user not found')
 
         let newpassword = await bycrypt.compare(password, userfind.password)
-
         if (!newpassword) throw new Error('password is not valid')
 
         const token = jwt.sign({ id: userfind._id }, "USER")
@@ -94,7 +114,7 @@ exports.READ = async function (req, res, next) {
                     { email: { $regex: req.query.search, $options: 'i' } },
                 ]
             });
-        } else{
+        } else {
             // No search query, fetch all data
             DATA = await NEWUSER.find();
         }
@@ -114,32 +134,62 @@ exports.READ = async function (req, res, next) {
 }
 
 
+
+
+
 exports.UPDATE = async function (req, res, next) {
     try {
 
-        const { firstname, lastname, email, password } = req.body;
+        let DATA;
+        const idcheck = await NEWUSER.findById(req.params.id)
+        if (!idcheck) throw new Error("Id is not exist");
 
-        let updatedPassword = password; 
+        const usercheck = await NEWUSER.findOne({ _id: req.user })
+        if (usercheck._id.toString() !== idcheck._id.toString()) throw new Error("UnauthorizedError: ID does not match");
 
-        if (password) {
-            updatedPassword = await bycrypt.hash(password, 8); 
+        let updateFields = {};
+
+        if (req.body.email) {
+            updateFields.email = req.body.email;
         }
-        
-        const DATA = await NEWUSER.findByIdAndUpdate(req.params.id, {
-            firstname: firstname,
-            lastname: lastname,
-            email: email,
-            password: updatedPassword
-        }, {
-            new: true // This should be inside the options object
-        });
 
+        if (req.body.password) {
+            updateFields.password = req.body.password;
+        }
 
-        res.status(201).json({
+        if (req.file && req.file.filename) {
+            updateFields.profile = req.file.filename;
+        }
+
+        if (req.file && req.file.filename) {
+            const directoryPath = path.join("..", 'CONATCT API', 'public', 'images');
+            const oldImagePath = path.join(directoryPath, idcheck.profile);
+
+            if (idcheck.profile && fs.existsSync(oldImagePath)) {
+                fs.unlink(oldImagePath, err => {
+                    if (err) console.error(`Error deleting old profile image:`, err);
+                    else console.log(`Deleted old profile image: ${idcheck.profile}`);
+                });
+
+                // let updatedPassword = password;
+                // if (password) {
+                //     updatedPassword = await bycrypt.hash(password, 8);
+                // }
+            }
+        }
+
+        if (Object.keys(updateFields).length > 0) {
+            DATA = await NEWUSER.findByIdAndUpdate(req.params.id, updateFields, { new: true });
+        } else {
+            throw new Error("No valid fields provided for update");
+        }
+
+        res.status(200).json({
             status: "Success",
             message: "Data Update Successfully",
-            data: DATA,
+            Updatedata: DATA,
         })
+
 
     } catch (error) {
         res.status(404).json({
@@ -148,24 +198,40 @@ exports.UPDATE = async function (req, res, next) {
         })
     }
 }
-
-
 
 exports.DELETE = async function (req, res, next) {
     try {
 
-        const DATA = await NEWUSER.findByIdAndDelete(req.params.id);
+        let DATA;
 
-        res.status(201).json({
+        const idcheck = await NEWUSER.findById(req.params.id)
+        if (!idcheck) throw new Error("User not found");
+
+        const usercheck = await NEWUSER.findOne({ _id: req.user })
+        if (usercheck._id.toString() !== idcheck._id.toString()) throw new Error("UnauthorizedError: ID does not match");
+
+        const directoryPath = path.join("..", 'CONATCT API', 'public', 'images');
+        const oldImagePath = path.join(directoryPath, idcheck.profile);
+        if (idcheck.profile && fs.existsSync(oldImagePath)) {
+            fs.unlink(oldImagePath, err => {
+                if (err) console.error(`Error deleting old profile image:`, err);
+                else console.log(`Deleted old profile image: ${idcheck.profile}`);
+            });
+        }
+
+        DATA = await NEWUSER.findByIdAndDelete(req.params.id);
+
+        res.status(200).json({
             status: "Success",
             message: "Data Delete Successfully",
+            data: DATA,
         })
-
-    } catch (error) {
+    }
+    catch (error) {
         res.status(404).json({
             status: "Fail",
             message: error.message,
         })
     }
-}
 
+}
